requirements.txt
requests[socks]
beautifulsoup4
colorama
python-dotenv
